requirements.txt
pytesseract
pdf2image
opencv-python
pandas
Pillow
